[~TASK] FLOW3 (Persistence): Added a safeguard in Session->unregisterReconstitutedEntity(). Original-Commit-Hash: dca3dd21781b9763b35af9638faa384dabd8a7b7

+++ TYPO3.Flow/Classes/Persistence/Session.php
@@ -115,8 +115,10 @@ class Session {
 	 * @author Karsten Dambekalns <karsten@typo3.org>
 	 */
 	public function unregisterReconstitutedEntity($entity) {
-		$this->reconstitutedEntities->detach($entity);
-		unset($this->reconstitutedEntitiesData[$this->getIdentifierByObject($entity)]);
+		if ($this->reconstitutedEntities->contains($entity)) {
+			$this->reconstitutedEntities->detach($entity);
+			unset($this->reconstitutedEntitiesData[$this->getIdentifierByObject($entity)]);
+		}
 	}
 
 	/**
